requirements.txt
gradio>=4.0.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.0.0
torch>=2.0.0
transformers>=4.28.0
matplotlib>=3.6.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0
pytest>=7.3.1
opencv-python>=4.6.0